# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sensible",
        "title": "Sensible Query Spec",
        "contact": {
            "name": "sensiblequery",
            "url": "https://github.com/sensible-contract/sensiblequery",
            "email": "jiedohh@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/contract/swap-aggregate-amount/{codehash}/{genesis}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unique"
                ],
                "summary": "查询Swap合约在区块中的聚合Amount数据，以合约CodeHash+GenesisID来确认一种Swap",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Start Block Height",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "End Block Height, (0 to get till latest)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 6,
                        "description": "每批聚合区块数量",
                        "name": "interval",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "844c56bb99afc374967a27ce3b46244e2e1fba60",
                        "description": "Code Hash160",
                        "name": "codehash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "74967a27ce3b46244e2e1fba60844c56bb99afc3",
                        "description": "Genesis ID",
                        "name": "genesis",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ContractSwapAggregateAmountResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/contract/swap-aggregate/{codehash}/{genesis}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unique"
                ],
                "summary": "查询Swap合约在区块中的聚合价格数据，以合约CodeHash+GenesisID来确认一种Swap",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Start Block Height",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "End Block Height, (0 to get till latest)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 6,
                        "description": "每批聚合区块数量",
                        "name": "interval",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "844c56bb99afc374967a27ce3b46244e2e1fba60",
                        "description": "Code Hash160",
                        "name": "codehash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "74967a27ce3b46244e2e1fba60844c56bb99afc3",
                        "description": "Genesis ID",
                        "name": "genesis",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ContractSwapAggregateResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/contract/swap-data/{codehash}/{genesis}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unique"
                ],
                "summary": "查询Swap合约在区块中的每次交易数据，以合约CodeHash+GenesisID来确认一种Swap",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Start Block Height",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "End Block Height, (0 to get mempool data)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始游标",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 16,
                        "description": "返回记录数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "844c56bb99afc374967a27ce3b46244e2e1fba60",
                        "description": "Code Hash160",
                        "name": "codehash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "74967a27ce3b46244e2e1fba60844c56bb99afc3",
                        "description": "Genesis ID",
                        "name": "genesis",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ContractSwapDataResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/token/info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "查询Token简述",
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": {}, \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TokenInfoResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ContractSwapAggregateAmountResp": {
            "type": "object",
            "properties": {
                "closeAmount": {
                    "description": "收盘Token1存量",
                    "type": "integer"
                },
                "height": {
                    "description": "区块高度",
                    "type": "integer"
                },
                "maxAmount": {
                    "description": "最高Token1存量",
                    "type": "integer"
                },
                "minAmount": {
                    "description": "最低Token1存量",
                    "type": "integer"
                },
                "openAmount": {
                    "description": "开盘Token1存量",
                    "type": "integer"
                },
                "timestamp": {
                    "description": "区块时间戳",
                    "type": "integer"
                },
                "txCount": {
                    "description": "交易次数",
                    "type": "integer"
                }
            }
        },
        "model.ContractSwapAggregateResp": {
            "type": "object",
            "properties": {
                "closePrice": {
                    "description": "收盘价格",
                    "type": "number"
                },
                "height": {
                    "description": "区块高度",
                    "type": "integer"
                },
                "maxPrice": {
                    "description": "最高价格",
                    "type": "number"
                },
                "minPrice": {
                    "description": "最低价格",
                    "type": "number"
                },
                "openPrice": {
                    "description": "开盘价格",
                    "type": "number"
                },
                "timestamp": {
                    "description": "区块时间戳",
                    "type": "integer"
                },
                "token1Volume": {
                    "description": "token1交易量",
                    "type": "integer"
                },
                "token2Volume": {
                    "description": "token2交易量",
                    "type": "integer"
                }
            }
        },
        "model.ContractSwapDataResp": {
            "type": "object",
            "properties": {
                "codeType": {
                    "description": "合约类型 0: None, 1: FT, 2: Unique, 3: NFT",
                    "type": "integer"
                },
                "height": {
                    "description": "区块高度",
                    "type": "integer"
                },
                "idx": {
                    "type": "integer"
                },
                "inToken1Amount": {
                    "type": "integer"
                },
                "inToken2Amount": {
                    "type": "integer"
                },
                "inTokenLpAmount": {
                    "type": "integer"
                },
                "operation": {
                    "description": "0: sell, 1: buy, 2: add, 3: remove",
                    "type": "integer"
                },
                "outToken1Amount": {
                    "type": "integer"
                },
                "outToken2Amount": {
                    "type": "integer"
                },
                "outTokenLpAmount": {
                    "type": "integer"
                },
                "timestamp": {
                    "description": "区块时间戳",
                    "type": "integer"
                },
                "txid": {
                    "description": "当前txid",
                    "type": "string"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.TokenInfoResp": {
            "type": "object",
            "properties": {
                "countCode": {
                    "description": "当前合约代码种类数",
                    "type": "integer"
                },
                "countFT": {
                    "description": "当前FT总数",
                    "type": "integer"
                },
                "countNFT": {
                    "description": "当前NFT总数",
                    "type": "integer"
                },
                "countNFTID": {
                    "description": "当前NFTID总数",
                    "type": "integer"
                },
                "countOwnerFT": {
                    "description": "当前持有FT人数",
                    "type": "integer"
                },
                "countOwnerNFT": {
                    "description": "当前持有NFT人数",
                    "type": "integer"
                },
                "inTimesFT": {
                    "description": "FT总输入次数",
                    "type": "integer"
                },
                "inTimesNFT": {
                    "description": "NFT总输入次数",
                    "type": "integer"
                },
                "outTimesFT": {
                    "description": "FT总输出次数",
                    "type": "integer"
                },
                "outTimesNFT": {
                    "description": "NFT总输出次数",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}